[ValidCheck][Refactor] ShellStringParser .cpp -> .h 파일 분기
