+++ Project1/Project1.vcxproj.filters
@@ -47,5 +47,8 @@
     <ClInclude Include="Invoker.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
+    <ClInclude Include="ShellStringParser.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
