+++ Project1/ShellStringParser.cpp
@@ -1,83 +1,66 @@
-#include <iostream>
-#include <string>
-#include <regex>
-
-using namespace std;
-
-enum class ShellStringParserError {
-	NO_ERROR,
-	CMD_EMPTY,
-	CMD_NOT_FOUND,
-	CMD_ARGC_ERROR,
-	CMD_ARGV_ERROR
-};
+#include "ShellStringParser.h"
 
 
-class ShellStringParser {
-public:
+ShellStringParserError ShellStringParser::validCheck(string inputCmd) {
+	// Error
+	if (inputCmd == "")
+		return ShellStringParserError::CMD_EMPTY;
 
-	ShellStringParserError validCheck(string inputCmd) {
-		// Error
-		if (inputCmd == "")
-			return ShellStringParserError::CMD_EMPTY;
-		
-		vector<string> cmdVec = split(inputCmd);
-		if (cmdVec[0] != "W" && cmdVec[0] != "R")
-		{
-			return ShellStringParserError::CMD_NOT_FOUND;
-		}
-		if (cmdVec[0] == "W")
-		{
-			if (cmdVec.size() != 3)
-				return ShellStringParserError::CMD_ARGC_ERROR;
+	vector<string> cmdVec = split(inputCmd);
+	if (cmdVec[0] != "W" && cmdVec[0] != "R")
+	{
+		return ShellStringParserError::CMD_NOT_FOUND;
+	}
+	if (cmdVec[0] == "W")
+	{
+		if (cmdVec.size() != 3)
+			return ShellStringParserError::CMD_ARGC_ERROR;
 
-			if (isNotLBA(cmdVec[1]))
-				return ShellStringParserError::CMD_ARGV_ERROR;
+		if (isNotLBA(cmdVec[1]))
+			return ShellStringParserError::CMD_ARGV_ERROR;
 
-			if (isNotValue(cmdVec[2]))
-				return ShellStringParserError::CMD_ARGV_ERROR;
-		}
-
-		return ShellStringParserError::NO_ERROR;
+		if (isNotValue(cmdVec[2]))
+			return ShellStringParserError::CMD_ARGV_ERROR;
 	}
 
-	bool isNotLBA(string strLBA) {
-		regex txt_regex("^[0-9]+$");
-		if (regex_match(strLBA, txt_regex) == false)
-			return true;
+	return ShellStringParserError::NO_ERROR;
+}
 
-		int nLBA = atoi(strLBA.c_str());
-		if (nLBA < 0 || nLBA > 100)
-			return true;
+bool ShellStringParser::isNotLBA(string strLBA) {
+	regex txt_regex("^[0-9]+$");
+	if (regex_match(strLBA, txt_regex) == false)
+		return true;
 
-		return false;
-	}
+	int nLBA = atoi(strLBA.c_str());
+	if (nLBA < 0 || nLBA > 100)
+		return true;
 
-	bool isNotValue(string strValue) {
-		regex txt_regex("^0x[0-9A-F]{8}$");
-		if (regex_match(strValue, txt_regex) == false)
-			return true;
+	return false;
+}
 
-		return false;
-	}
-	
-	vector<string> split(string inputCmd) {
-		vector<string> cmdVector;
-		int nPrevPos = 0;
-		int nNextPos = inputCmd.find(" ");
-		string str = "";
-		while (nNextPos != string::npos) {
-			str = inputCmd.substr(nPrevPos, nNextPos - nPrevPos);
-			if (str != "")
-				cmdVector.push_back(str);
-			nPrevPos = nNextPos + 1;
-			nNextPos = inputCmd.find(" ", nNextPos + 1);
-		}
+bool ShellStringParser::isNotValue(string strValue) {
+	regex txt_regex("^0x[0-9A-F]{8}$");
+	if (regex_match(strValue, txt_regex) == false)
+		return true;
+
+	return false;
+}
+
+vector<string> ShellStringParser::split(string inputCmd) {
+	vector<string> cmdVector;
+	int nPrevPos = 0;
+	int nNextPos = inputCmd.find(" ");
+	string str = "";
+	while (nNextPos != string::npos) {
 		str = inputCmd.substr(nPrevPos, nNextPos - nPrevPos);
 		if (str != "")
 			cmdVector.push_back(str);
-
-		return cmdVector;
+		nPrevPos = nNextPos + 1;
+		nNextPos = inputCmd.find(" ", nNextPos + 1);
 	}
-};
+	str = inputCmd.substr(nPrevPos, nNextPos - nPrevPos);
+	if (str != "")
+		cmdVector.push_back(str);
 
+	return cmdVector;
+}
