+++ Project1/ShellStringParser.h
@@ -1,0 +1,26 @@
+#pragma once
+#include <iostream>
+#include <string>
+#include <regex>
+
+using namespace std;
+
+enum class ShellStringParserError {
+	NO_ERROR,
+	CMD_EMPTY,
+	CMD_NOT_FOUND,
+	CMD_ARGC_ERROR,
+	CMD_ARGV_ERROR
+};
+
+
+class ShellStringParser {
+public:
+	ShellStringParserError validCheck(string inputCmd);
+
+private:
+	bool isNotLBA(string strLBA);
+	bool isNotValue(string strValue);
+	vector<string> split(string inputCmd);
+};
+
